+++ Assets~/StandardAssets/Prefabs/XRCameraRig.prefab
@@ -11,7 +11,7 @@ GameObject:
   - component: {fileID: 6074567058768481978}
   - component: {fileID: 8828591944363462759}
   m_Layer: 0
-  m_Name: TrackedCameraRig
+  m_Name: XRCameraRig
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
