+++ Assets~/StandardAssets/Prefabs/XRCameraRig.prefab.meta
