+++ Assets~/StandardAssets/Prefabs/XRPlayerController.prefab
@@ -205,6 +205,7 @@ MonoBehaviour:
   bodyDiameter: 1
   startRotateBodyAngle: 45
   angleRotateBodySpeed: 20
+  gravityMode: 0
 --- !u!54 &4329863972925498748
 Rigidbody:
   m_ObjectHideFlags: 0
@@ -238,7 +239,7 @@ CharacterController:
   m_StepOffset: 0.27
   m_SkinWidth: 0.05
   m_MinMoveDistance: 0.001
-  m_Center: {x: 0, y: 0.675, z: 0}
+  m_Center: {x: 0, y: 0.725, z: 0}
 --- !u!1 &9104542960344264379
 GameObject:
   m_ObjectHideFlags: 0
