+++ Runtime/Interfaces/IXRPlayerController.cs
@@ -1,14 +1,15 @@
 // Copyright (c) Reality Collective. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
+using RealityToolkit.CameraService.Definitions;
 using UnityEngine;
 
 namespace RealityToolkit.CameraService.Interfaces
 {
     /// <summary>
-    /// A <see cref="ICameraRig"/> for player character simulation.
+    /// A <see cref="IXRCameraRig"/> for player character simulation.
     /// </summary>
-    public interface IPlayerCameraRig : ICameraRig
+    public interface IXRPlayerController : IXRCameraRig
     {
         /// <summary>
         /// Controls when gravity begins to take effect.
